IDVA6-2271 More tests refactored
