+++ test/integration/your-companies/confirmation-company-added/get.test.ts
@@ -11,7 +11,7 @@ import * as enCommon from "../../../../locales/en/common.json";
 import * as cyCommon from "../../../../locales/cy/common.json";
 
 const router = supertest(app);
-const url = `/your-companies/confirmation-company-added`;
+const url = "/your-companies/confirmation-company-added";
 const companyNumber = "1122334455";
 const companyName = "Acme Ltd";
 const session: Session = new Session();
@@ -22,15 +22,14 @@ mocks.mockSessionMiddleware.mockImplementation((req: Request, res: Response, nex
     return next();
 });
 
-describe(`GET ${url}`, () => {
+const redirectPageSpy: jest.SpyInstance = jest.spyOn(referrerUtils, "redirectPage");
 
-    const redirectPageSpy: jest.SpyInstance = jest.spyOn(referrerUtils, "redirectPage");
+describe("GET /your-companies/confirmation-company-added", () => {
 
     beforeEach(() => {
         jest.clearAllMocks();
+        redirectPageSpy.mockReturnValue(false);
     });
-
-    redirectPageSpy.mockReturnValue(false);
 
     it("should check session, auth and company authorisation before returning the your-companies page", async () => {
         await router.get(url);
@@ -38,48 +37,41 @@ describe(`GET ${url}`, () => {
         expect(mocks.mockEnsureSessionCookiePresentMiddleware).toHaveBeenCalled();
         expect(mocks.mockAuthenticationMiddleware).toHaveBeenCalled();
     });
-    it("should return status 200", async () => {
-        await router.get(url).expect(200);
-    });
 
-    it("should return expected English content if no language selected", async () => {
-        const response = await router.get(`${url}`);
-        expect(response.text).toContain(enCommon.success);
-    });
-
-    it("should display the company name", async () => {
-        const response = await router.get(`${url}`);
-        expect(response.text).toContain("Acme Ltd");
-    });
-
-    it("should display 3 bullet points in English if language set to English", async () => {
-        const response = await router.get(`${url}?lang=en`);
-        expect(response.text).toContain(en.bullet_1);
-        expect(response.text).toContain(en.bullet_2);
-        expect(response.text).toContain(en.bullet_3);
-    });
-
-    it("should display 3 bullet points in Welsh if language set to Welsh", async () => {
-        const response = await router.get(`${url}?lang=cy`);
-        expect(response.text).toContain(cy.bullet_1);
-        expect(response.text).toContain(cy.bullet_2);
-        expect(response.text).toContain(cy.bullet_3);
-    });
-
-    it("should return expected Welsh content when welsh is selected", async () => {
-        const response = await router.get(`${url}?lang=cy`);
-        expect(response.text).toContain(cyCommon.success);
-    });
+    test.each([
+        // Given
+        { langInfo: "English", langVersion: "en", lang: en, langCommon: enCommon },
+        { langInfo: "English", langVersion: undefined, lang: en, langCommon: enCommon },
+        { langInfo: "Welsh", langVersion: "cy", lang: cy, langCommon: cyCommon }
+    ])("should return status 200 and expected $langInfo content if language set to $langVersion",
+        async ({ langVersion, lang, langCommon }) => {
+            // When
+            const response = await router.get(`${url}?lang=${langVersion}`);
+            // Then
+            expect(response.status).toEqual(200);
+            expect(response.text).toContain(companyName);
+            expect(response.text).toContain(langCommon.success);
+            expect(response.text).toContain(lang.bullet_1);
+            expect(response.text).toContain(lang.bullet_2);
+            expect(response.text).toContain(lang.bullet_3);
+        });
 
     it("should return status 302 on page redirect", async () => {
+        // Given
         redirectPageSpy.mockReturnValue(true);
-        await router.get(url).expect(302);
+        // When
+        const response = await router.get(url);
+        // Then
+        expect(response.status).toEqual(302);
     });
 
     it("should return correct response message including desired url path", async () => {
+        // Given
         const urlPath = LANDING_URL;
         redirectPageSpy.mockReturnValue(true);
+        // When
         const response = await router.get(url);
+        // Then
         expect(response.text).toEqual(`Found. Redirecting to ${urlPath}`);
     });
 });
